Move settings_controller.py into RimSort.controllers
